+++ Runtime/Scripts/GlobalTools/Types/TManager.cs
@@ -3,10 +3,11 @@ using System.Collections.Generic;
 using UnityEngine;
 
 namespace PolytopeSolutions.Toolset.GlobalTools.Types {
-	public class TManager<TItem> : MonoBehaviour
-		                         where TItem : MonoBehaviour {
+	public abstract class TManager<TItem> : MonoBehaviour
+											where TItem : MonoBehaviour {
+		// NB! if overriding Awake - call base.Awake() in derived classes.
 		private static TItem _instance;
-		public static TItem instance {
+		public static TItem Instance {
 			get {
 				if (_instance == null) {
 					// Find It
@@ -14,6 +15,10 @@ namespace PolytopeSolutions.Toolset.GlobalTools.Types {
 				}
 				return _instance;
 			}
+		}
+		protected virtual void Awake() { 
+			if (!Instance)
+				Debug.LogWarning("No instance of " + typeof(TItem).Name + " found in scene.");
 		}
 	}
 }