+++ Runtime/Scripts/Input/InputReceiver.cs
@@ -1,23 +1,26 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-
+using UnityEngine.EventSystems;
+using UnityEngine.InputSystem;
 
 namespace PolytopeSolutions.Toolset.Input {
     public abstract class InputReceiver : MonoBehaviour {
         [Header("Input Receiver Settings")]
         [SerializeField] protected bool isActiveByDefault;
         [SerializeField] protected string inputReceiverKeyName;
-        protected bool isInputEnabled = false;
+        protected bool isReceiverEnabled = false;
+        protected bool isTemporaryDisabled = false;
+        protected bool IsInputEnabled => this.isReceiverEnabled && !this.isTemporaryDisabled;
         public bool IsActiveByDefault => this.isActiveByDefault;
-        private bool isSelfManaged => InputManager.instance == null;
+        private bool IsSelfManaged => InputManager.Instance == null;
 
         protected abstract void EnableInputEvents();
         protected abstract void DisableInputEvents();
         public virtual void SetActiveInputs(bool targetState) {
-            if (this.isInputEnabled == targetState) return;
-            this.isInputEnabled = targetState;
-            if (this.isInputEnabled)
+            if (this.isReceiverEnabled == targetState) return;
+            this.isReceiverEnabled = targetState;
+            if (this.isReceiverEnabled)
                 EnableInputEvents();
             else
                 DisableInputEvents();
@@ -26,15 +29,15 @@ namespace PolytopeSolutions.Toolset.Input {
         protected virtual void Awake() {
             if (string.IsNullOrEmpty(this.inputReceiverKeyName))
                 this.inputReceiverKeyName = gameObject.name+"_InputReceiver";
-            if (!this.isSelfManaged)
-                InputManager.instance.RegisterInputReceiver(this.inputReceiverKeyName, this);
+            if (!this.IsSelfManaged)
+                InputManager.Instance.RegisterInputReceiver(this.inputReceiverKeyName, this);
         }
         protected virtual void OnEnable() { 
-            if (this.isSelfManaged)
+            if (this.IsSelfManaged)
                 SetActiveInputs(this.IsActiveByDefault);
         }
         protected virtual void OnDisable() { 
-            if (this.isSelfManaged)
+            if (this.IsSelfManaged)
                 SetActiveInputs(false);
         }
     }