+++ Runtime/Scripts/Scenes/LoadingBar.cs
@@ -6,6 +6,6 @@ using PolytopeSolutions.Toolset.GlobalTools.UI;
 
 namespace PolytopeSolutions.Toolset.Scenes {
     public class LoadingBar : ProgressBar {
-        protected override float targetProgress => (float)SceneManagerExtender.instance?.Progress;
+        protected override float targetProgress => (float)SceneManagerExtender.Instance?.LoadingProgress;
     }
 }